refactor install func to accept custom package name
