+++ src/commands/__init__.py
@@ -1,2 +1,2 @@
-from src.commands.install import install_package, installation_exceptions
+from src.commands.install import handle_install_command
 from src.commands.remove import removal_exceptions, remove
