--- src/commands/install/utils/__init__.py
@@ -1,3 +1,0 @@
-from src.commands.install.utils.extract_info_from_repo_id import (
-    extract_info_from_repo_id,
-)
